revise events-to-ts conversion algorithm
the previous algorithm only adds a value if the time tick happen to be
inside a event. this would cause it to miss event scores in between
ticks.

in this revision, we update the conversion algorithm to calculate the
max score that overlaps with each given time range (tick + step), and
thus will not miss any events that happened during a range.